# src/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.Modality": {
				"type": "string",
				"enum": [
					"Onsite",
					"Remote",
					"Hybrid"
				]
			},
			"Modality": {
				"$ref": "#/components/schemas/_36_Enums.Modality"
			},
			"_36_Enums.InternshipType": {
				"type": "string",
				"enum": [
					"Curricular",
					"Extracurricular"
				]
			},
			"InternshipType": {
				"$ref": "#/components/schemas/_36_Enums.InternshipType"
			},
			"_36_Enums.Workday": {
				"type": "string",
				"enum": [
					"FullTime",
					"PartTime"
				]
			},
			"Workday": {
				"$ref": "#/components/schemas/_36_Enums.Workday"
			},
			"Degree": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Internship": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"internshipTitle": {
						"type": "string"
					},
					"internshipLocation": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"modality": {
						"$ref": "#/components/schemas/Modality"
					},
					"internshipType": {
						"$ref": "#/components/schemas/InternshipType"
					},
					"workday": {
						"$ref": "#/components/schemas/Workday"
					},
					"minimumStudies": {
						"type": "string"
					},
					"languages": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"period": {
						"type": "string"
					},
					"minimumExperience": {
						"type": "string"
					},
					"backgroundKnowledge": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"degree": {
						"$ref": "#/components/schemas/Degree"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					}
				},
				"required": [
					"id",
					"internshipTitle",
					"internshipLocation",
					"salary",
					"modality",
					"internshipType",
					"workday",
					"minimumStudies",
					"languages",
					"startDate",
					"period",
					"minimumExperience",
					"backgroundKnowledge",
					"description",
					"degree",
					"company"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternshipInput": {
				"properties": {
					"internshipTitle": {
						"type": "string"
					},
					"internshipLocation": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"modality": {
						"$ref": "#/components/schemas/Modality"
					},
					"internshipType": {
						"$ref": "#/components/schemas/InternshipType"
					},
					"workday": {
						"$ref": "#/components/schemas/Workday"
					},
					"minimumStudies": {
						"type": "string"
					},
					"languages": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"period": {
						"type": "string"
					},
					"minimumExperience": {
						"type": "string"
					},
					"backgroundKnowledge": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"degree_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"company_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"internshipTitle",
					"internshipLocation",
					"salary",
					"modality",
					"internshipType",
					"workday",
					"minimumStudies",
					"languages",
					"startDate",
					"period",
					"minimumExperience",
					"backgroundKnowledge",
					"description",
					"degree_id",
					"company_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SafeUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/SafeUser"
					}
				},
				"required": [
					"token",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthInput": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterInput": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "cemubackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/internships/{id}": {
			"get": {
				"operationId": "GetInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Internship"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID"
					},
					"404": {
						"description": "Internship not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Internship"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID"
					},
					"404": {
						"description": "Internship not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternshipInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID"
					},
					"404": {
						"description": "Internship not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/internships": {
			"get": {
				"operationId": "GetInternships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Internship"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostInternship",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Internship"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternshipInput"
							}
						}
					}
				}
			}
		},
		"/degrees": {
			"get": {
				"operationId": "GetDegrees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Degree"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Degrees"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateDegree",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Degree"
								}
							}
						}
					}
				},
				"tags": [
					"Degrees"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Degree"
							}
						}
					}
				}
			}
		},
		"/degrees/{id}": {
			"get": {
				"operationId": "GetDegree",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Degree"
								}
							}
						}
					},
					"404": {
						"description": "Degree not found"
					}
				},
				"tags": [
					"Degrees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDegree",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Degree"
								}
							}
						}
					},
					"404": {
						"description": "Degree not found"
					}
				},
				"tags": [
					"Degrees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Degree"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDegree",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Degree not found"
					}
				},
				"tags": [
					"Degrees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/companies": {
			"get": {
				"operationId": "GetCompanies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Company"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostCompany",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			}
		},
		"/companies/{id}": {
			"get": {
				"operationId": "GetCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					},
					"404": {
						"description": "Company not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					},
					"404": {
						"description": "Company not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Company not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid credentials"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthInput"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "User registered",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid registration data"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterInput"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}